feat(format): change print width to 120
This is against prettier's recommendation, but there are lots of cases
where 80 is just too short. I hard disagree that things spread on
multiple lines is always a readability gain and this will give more
flexibility in that. For context, I thought 120 was the default.

BREAKING CHANGE: will change the way things are formatted

Closes #34